+++ miden-tx/src/testing/tx_context/builder.rs
@@ -52,7 +52,7 @@ pub struct TransactionContextBuilder {
 impl TransactionContextBuilder {
     pub fn new(account: Account) -> Self {
         Self {
-            assembler: TransactionKernel::testing_assembler(),
+            assembler: TransactionKernel::testing_assembler_with_mock_account(),
             account,
             account_seed: None,
             input_notes: Vec::new(),