+++ objects/src/accounts/code/procedure.rs
@@ -1,3 +1,5 @@
+use alloc::string::ToString;
+
 use vm_core::{
     utils::{ByteReader, ByteWriter, Deserializable, Serializable},
     FieldElement,
@@ -5,40 +7,80 @@ use vm_core::{
 use vm_processor::DeserializationError;
 
 use super::{Digest, Felt};
-use crate::AccountError;
+use crate::{accounts::AccountStorage, AccountError};
 
 // ACCOUNT PROCEDURE INFO
 // ================================================================================================
 
 /// Information about a procedure exposed in a public account interface.
 ///
-/// The info included the MAST root of the procedure and the storage offset applied to all account
-/// storage-related accesses made by this procedure. For example, if storage offset is set ot 1, a
-/// call to the account::get_item(storage_slot=4) made from this procedure would actually access
+/// The info included the MAST root of the procedure, the storage offset applied to all account
+/// storage-related accesses made by this procedure and the storage size allowed to be accessed
+/// by this procedure.
+///
+/// The offset is applied to any accesses made from within the procedure to the associated
+/// account's storage. For example, if storage offset for a procedure is set ot 1, a call
+/// to the account::get_item(storage_slot=4) made from this procedure would actually access
 /// storage slot with index 5.
+///
+/// The size is used to limit how many storage slots a given procedure can access in the associated
+/// account's storage. For example, if storage size for a procedure is set to 3, the procedure will
+/// be bounded to access storage slots in the range [storage_offset, storage_offset + 3 - 1].
+/// Furthermore storage_size = 0 indicates that a procedure does not need to access storage.
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct AccountProcedureInfo {
     mast_root: Digest,
-    storage_offset: u16,
+    storage_offset: u8,
+    storage_size: u8,
 }
 
 impl AccountProcedureInfo {
     /// The number of field elements needed to represent an [AccountProcedureInfo] in kernel memory.
     pub const NUM_ELEMENTS_PER_PROC: usize = 8;
 
+    // CONSTRUCTOR
+    // --------------------------------------------------------------------------------------------
+
     /// Returns a new instance of an [AccountProcedureInfo].
-    pub fn new(mast_root: Digest, storage_offset: u16) -> Self {
-        Self { mast_root, storage_offset }
+    ///
+    /// # Errors
+    /// - If `storage_size` is 0 and `storage_offset` is not 0.
+    /// - If `storage_size + storage_offset` is greater than `MAX_NUM_STORAGE_SLOTS`.
+    pub fn new(
+        mast_root: Digest,
+        storage_offset: u8,
+        storage_size: u8,
+    ) -> Result<Self, AccountError> {
+        if storage_size == 0 && storage_offset != 0 {
+            return Err(AccountError::PureProcedureWithStorageOffset);
+        }
+
+        if (storage_offset + storage_size) as usize > AccountStorage::MAX_NUM_STORAGE_SLOTS {
+            return Err(AccountError::StorageOffsetOutOfBounds {
+                max: AccountStorage::MAX_NUM_STORAGE_SLOTS as u8,
+                actual: storage_offset + storage_size,
+            });
+        }
+
+        Ok(Self { mast_root, storage_offset, storage_size })
     }
 
-    /// Returns a reference to the procedure's mast_root.
+    // PUBLIC ACCESSORS
+    // --------------------------------------------------------------------------------------------
+
+    /// Returns a reference to the procedure's mast root.
     pub fn mast_root(&self) -> &Digest {
         &self.mast_root
     }
 
-    /// Returns a reference to the procedure's storage_offset.
-    pub fn storage_offset(&self) -> u16 {
+    /// Returns the procedure's storage offset.
+    pub fn storage_offset(&self) -> u8 {
         self.storage_offset
+    }
+
+    /// Returns the procedure's storage size.
+    pub fn storage_size(&self) -> u8 {
+        self.storage_size
     }
 }
 
@@ -50,7 +92,10 @@ impl From<AccountProcedureInfo> for [Felt; 8] {
         result[0..4].copy_from_slice(value.mast_root().as_elements());
 
         // copy the storage offset into value[4]
-        result[4] = Felt::from(value.storage_offset());
+        result[4] = Felt::from(value.storage_offset);
+
+        // copy the storage size into value[7]
+        result[5] = Felt::from(value.storage_size);
 
         result
     }
@@ -64,36 +109,45 @@ impl TryFrom<[Felt; 8]> for AccountProcedureInfo {
         let mast_root = Digest::from(<[Felt; 4]>::try_from(&value[0..4]).unwrap());
 
         // get storage_offset form value[4]
-        let storage_offset: u16 = value[4]
+        let storage_offset: u8 = value[4]
             .try_into()
             .map_err(|_| AccountError::AccountCodeProcedureInvalidStorageOffset)?;
 
-        // Check if the last three elements are zero
-        if value[5..].iter().any(|&x| x != Felt::ZERO) {
+        // get storage_size form value[5]
+        let storage_size: u8 = value[5]
+            .try_into()
+            .map_err(|_| AccountError::AccountCodeProcedureInvalidStorageSize)?;
+
+        // Check if the remaining values are 0
+        if value[6] != Felt::ZERO || value[7] != Felt::ZERO {
             return Err(AccountError::AccountCodeProcedureInvalidPadding);
         }
 
-        Ok(Self { mast_root, storage_offset })
+        Ok(Self { mast_root, storage_offset, storage_size })
     }
 }
 
 impl Serializable for AccountProcedureInfo {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        target.write(self.mast_root());
-        target.write_u16(self.storage_offset());
+        target.write(self.mast_root);
+        target.write_u8(self.storage_offset);
+        target.write_u8(self.storage_size)
     }
 
     fn get_size_hint(&self) -> usize {
-        self.mast_root.get_size_hint() + self.storage_offset.get_size_hint()
+        self.mast_root.get_size_hint()
+            + self.storage_offset.get_size_hint()
+            + self.storage_size.get_size_hint()
     }
 }
 
 impl Deserializable for AccountProcedureInfo {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let mast_root: Digest = source.read()?;
-        let storage_offset = source.read_u16()?;
-
-        Ok(Self::new(mast_root, storage_offset))
+        let storage_offset = source.read_u8()?;
+        let storage_size = source.read_u8()?;
+        Self::new(mast_root, storage_offset, storage_size)
+            .map_err(|err| DeserializationError::InvalidValue(err.to_string()))
     }
 }
 
@@ -102,9 +156,26 @@ impl Deserializable for AccountProcedureInfo {
 
 #[cfg(test)]
 mod tests {
+
     use miden_crypto::utils::{Deserializable, Serializable};
+    use vm_core::Felt;
 
     use crate::accounts::{AccountCode, AccountProcedureInfo};
+
+    #[test]
+    fn test_from_to_account_procedure() {
+        let account_code = AccountCode::mock();
+
+        let procedure = account_code.procedures()[0].clone();
+
+        // from procedure to [Felt; 8]
+        let felts: [Felt; 8] = procedure.clone().into();
+
+        // try_from [Felt; 8] to procedure
+        let final_procedure: AccountProcedureInfo = felts.try_into().unwrap();
+
+        assert_eq!(procedure, final_procedure);
+    }
 
     #[test]
     fn test_serde_account_procedure() {
@@ -113,6 +184,6 @@ mod tests {
         let serialized = account_code.procedures()[0].to_bytes();
         let deserialized = AccountProcedureInfo::read_from_bytes(&serialized).unwrap();
 
-        assert_eq!(deserialized, account_code.procedures()[0]);
+        assert_eq!(account_code.procedures()[0], deserialized);
     }
 }